refactor(fixed periods): restructure folders and names
