+++ src/period-calculation/daily-periods/build-daily-fixed-period.ts
@@ -1,7 +1,7 @@
 import { Temporal } from '@js-temporal/polyfill'
 import { formatYyyyMmDD } from '../../utils/helpers'
 
-const buildDailyPeriod = ({ date }: { date: Temporal.PlainDate }) => {
+const buildDailyFixedPeriod = ({ date }: { date: Temporal.PlainDate }) => {
     const year = date.year
     const nextDayMonthLabel = String(date.month).padStart(2, '0')
     const nextDayLabel = String(date.day).padStart(2, '0')
@@ -16,4 +16,4 @@ const buildDailyPeriod = ({ date }: { date: Temporal.PlainDate }) => {
     }
 }
 
-export default buildDailyPeriod
+export default buildDailyFixedPeriod
