+++ src/period-calculation/daily-periods/index.ts
@@ -1,2 +1,1 @@
-export { default as getDailyPeriods } from './get-daily-periods'
-export { default as getDailyPeriodByDate } from './get-daily-period-by-date'
+export { default as buildDailyFixedPeriod } from './build-daily-fixed-period'
