+++ src/period-calculation/generate-fixed-periods/generate-fixed-periods.gregorian.spec.ts
@@ -1,74 +1,87 @@
-import { SupportedCalendar } from '../types'
-import { generateFixedPeriods } from './fixed-periods'
+import { SupportedCalendar } from '../../types'
+import generateFixedPeriods from './generate-fixed-periods'
 
-describe('Ethiopic Calendar fixed period calculation', () => {
+describe('Gregorian Calendar fixed period calculation', () => {
     describe('weekly periods', () => {
-        it('should respect starting day', () => {
-            const resultStartingMonday = generateFixedPeriods({
+        it('should omit last week if most of it is in next year', () => {
+            const results = generateFixedPeriods({
                 periodType: 'WEEKLY',
-                year: 2015,
-                calendar: 'ethiopic',
+                year: 2014,
+                calendar: 'gregory',
                 locale: 'en',
                 startingDay: 1,
             })
-            const resultStartingTuesday = generateFixedPeriods({
-                periodType: 'WEEKLY',
-                year: 2015,
-                calendar: 'ethiopic',
-                locale: 'en',
-                startingDay: 2,
+            expect(results[results.length - 1]).toMatchObject({
+                id: '2014W52',
+                iso: '2014W52',
+                name: 'Week 52 - 2014-12-22 - 2014-12-28',
             })
-
-            expect(resultStartingMonday[0]).not.toEqual(
-                resultStartingTuesday[0]
-            )
+        })
+        it('should start the year before if necessary', () => {
+            const results = generateFixedPeriods({
+                periodType: 'WEEKLYSUN',
+                year: 2014,
+                calendar: 'gregory',
+                locale: 'en',
+                startingDay: 1,
+            })
+            expect(results[0]).toMatchObject({
+                id: '2014SunW1',
+                iso: '2014SunW1',
+                name: 'Week 1 - 2013-12-29 - 2014-01-04',
+            })
+            expect(results[52]).toMatchObject({
+                id: '2014SunW53',
+                iso: '2014SunW53',
+                name: 'Week 53 - 2014-12-28 - 2015-01-03',
+            })
         })
     })
 
     describe('financial year', () => {
-        const date = {
+        const gregorianDate = {
             year: 2015,
-            calendar: 'ethiopic' as SupportedCalendar,
+            calendar: 'gregory' as SupportedCalendar,
             locale: 'en',
         }
         it('should build the label for Financial Year November properly', () => {
             const result = generateFixedPeriods({
+                ...gregorianDate,
                 periodType: 'FYNOV',
-                ...date,
             })
 
             expect(result[0]).toMatchObject({
                 id: '2015Nov',
-                name: 'Hamle 2015 - Sene 2016',
+                name: 'November 2015 - October 2016',
             })
 
             expect(result[result.length - 1]).toMatchObject({
                 id: '2006Nov',
-                name: 'Hamle 2006 - Sene 2007',
+                name: 'November 2006 - October 2007',
             })
         })
 
         it('should build the label for Financial Year April properly', () => {
             const result = generateFixedPeriods({
                 periodType: 'FYAPR',
-                ...date,
+                ...gregorianDate,
             })
 
             expect(result[0]).toMatchObject({
                 id: '2015April',
-                name: 'Tahsas 2015 - Hedar 2016',
+                name: 'April 2015 - March 2016',
             })
 
             expect(result[result.length - 1]).toMatchObject({
                 id: '2006April',
-                name: 'Tahsas 2006 - Hedar 2007',
+                name: 'April 2006 - March 2007',
             })
         })
 
         it('should build the label for Financial Year properly', () => {
             const result = generateFixedPeriods({
                 periodType: 'YEARLY',
-                ...date,
+                ...gregorianDate,
             })
 
             expect(result[0]).toMatchObject({
@@ -88,7 +101,7 @@ describe('Ethiopic Calendar fixed period calculation', () => {
     describe('start and end dates', () => {
         const date = {
             year: 2015,
-            calendar: 'ethiopic' as const,
+            calendar: 'gregory' as const,
             locale: 'en',
             startingDay: 1,
         }
@@ -105,18 +118,27 @@ describe('Ethiopic Calendar fixed period calculation', () => {
                 periodType: 'YEARLY',
                 ...date,
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods[0]).toEqual('2015-01-01/2015-13-06')
-            expect(periods[1]).toEqual('2014-01-01/2014-13-05')
-            expect(periods[periods.length - 1]).toEqual('2006-01-01/2006-13-05')
+            expect(periods[0]).toEqual('2015-01-01/2015-12-31')
+            expect(periods[1]).toEqual('2014-01-01/2014-12-31')
+            expect(periods[periods.length - 1]).toEqual('2006-01-01/2006-12-31')
         })
         it('should add start and end dates for FYNOV', () => {
             const periods = generateFixedPeriods({
                 ...date,
                 periodType: 'FYNOV',
-                year: 2015,
+                year: 2005,
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods[0]).toEqual('2015-11-01/2016-10-30')
-            expect(periods[periods.length - 1]).toEqual('2006-11-01/2007-10-30')
+            expect(periods[0]).toEqual('2005-11-01/2006-10-31')
+            expect(periods[periods.length - 1]).toEqual('1996-11-01/1997-10-31')
+        })
+        it('should add start and end dates for FYOCT', () => {
+            const periods = generateFixedPeriods({
+                ...date,
+                periodType: 'FYOCT',
+                year: 2005,
+            }).map((p) => `${p.startDate}/${p.endDate}`)
+            expect(periods[0]).toEqual('2005-10-01/2006-09-30')
+            expect(periods[periods.length - 1]).toEqual('1996-10-01/1997-09-30')
         })
         it('should add start and end dates for WEEKLY', () => {
             const periods = generateFixedPeriods({
@@ -124,18 +146,17 @@ describe('Ethiopic Calendar fixed period calculation', () => {
                 periodType: 'WEEKLY',
                 year: 2014,
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods[0]).toEqual('2014-01-04/2014-01-10')
-            expect(periods[periods.length - 1]).toEqual('2014-13-01/2015-01-02')
+            expect(periods[0]).toEqual('2013-12-30/2014-01-05')
+            expect(periods[periods.length - 1]).toEqual('2014-12-22/2014-12-28')
         })
         it('should add start and end dates for WEEKLYSUN', () => {
             const periods = generateFixedPeriods({
                 ...date,
-                periodType: 'WEEKLYWED',
+                periodType: 'WEEKLYSUN',
                 year: 2014,
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods.length).toEqual(52)
-            expect(periods[0]).toEqual('2013-13-04/2014-01-05') // starts from month Pagamen (13) as it should
-            expect(periods[periods.length - 1]).toEqual('2014-12-26/2014-13-02')
+            expect(periods[0]).toEqual('2013-12-29/2014-01-04')
+            expect(periods[periods.length - 1]).toEqual('2014-12-28/2015-01-03')
         })
         it('should add start and end dates for BIWEEKLY', () => {
             const periods = generateFixedPeriods({
@@ -143,21 +164,17 @@ describe('Ethiopic Calendar fixed period calculation', () => {
                 periodType: 'BIWEEKLY',
                 year: 2014,
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods.length).toEqual(26)
-            expect(periods[0]).toEqual('2014-01-04/2014-01-17')
-            expect(periods[periods.length - 1]).toEqual('2014-12-24/2015-01-02')
+            expect(periods[0]).toEqual('2013-12-30/2014-01-12')
+            expect(periods[periods.length - 1]).toEqual('2014-12-15/2014-12-28')
         })
         it('should add start and end dates for MONTHLY', () => {
             const periods = generateFixedPeriods({
                 ...date,
                 periodType: 'MONTHLY',
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods.length).toEqual(12) // 13th month is ignored
-            expect(periods[0]).toEqual('2015-01-01/2015-01-30')
-            expect(periods[1]).toEqual('2015-02-01/2015-02-30')
-            expect(periods[periods.length - 2]).toEqual('2015-11-01/2015-11-30')
-            // todo: check with Abyot what to do with the 13th month
-            expect(periods[periods.length - 1]).toEqual('2015-12-01/2015-12-30')
+            expect(periods[0]).toEqual('2015-01-01/2015-01-31')
+            expect(periods[1]).toEqual('2015-02-01/2015-02-28')
+            expect(periods[periods.length - 1]).toEqual('2015-12-01/2015-12-31')
         })
         it('should add start and end dates for SIXMONTHLY', () => {
             const periods = generateFixedPeriods({
@@ -165,36 +182,31 @@ describe('Ethiopic Calendar fixed period calculation', () => {
                 periodType: 'SIXMONTHLY',
             }).map((p) => `${p.startDate}/${p.endDate}`)
             expect(periods[0]).toEqual('2015-01-01/2015-06-30')
-            // todo: check with Abyot where the 13th month go in SIXMONTHLY
-            // expect(periods[periods.length - 1]).toEqual("2015-07-01/2015-12-30");
+            expect(periods[periods.length - 1]).toEqual('2015-07-01/2015-12-31')
         })
         it('should add start and end dates for SIXMONTHLYAPR', () => {
             const periods = generateFixedPeriods({
                 ...date,
                 periodType: 'SIXMONTHLYAPR',
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods.length).toEqual(2)
             expect(periods[0]).toEqual('2015-04-01/2015-09-30')
-            // todo: ignore the 13th month here ¬
-            // expect(periods[periods.length - 1]).toEqual("2015-10-01/2016-03-30");
+            expect(periods[periods.length - 1]).toEqual('2015-10-01/2016-03-31')
         })
         it('should add start and end dates for QUARTERLY', () => {
             const periods = generateFixedPeriods({
                 ...date,
                 periodType: 'QUARTERLY',
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods[0]).toEqual('2015-01-01/2015-03-30')
-            // todo: check with Abyot where the 13th month go in QUARTERLY
-            expect(periods[periods.length - 1]).toEqual('2015-10-01/2015-12-30')
+            expect(periods[0]).toEqual('2015-01-01/2015-03-31')
+            expect(periods[periods.length - 1]).toEqual('2015-10-01/2015-12-31')
         })
         it('should add start and end dates for BIMONTHLY', () => {
             const periods = generateFixedPeriods({
                 ...date,
                 periodType: 'BIMONTHLY',
             }).map((p) => `${p.startDate}/${p.endDate}`)
-            expect(periods[0]).toEqual('2015-01-01/2015-02-30')
-            // todo: check with Abyot if we want the 13th month as part of the end date
-            // expect(periods[periods.length - 1]).toEqual("2015-11-01/2015-13-06");
+            expect(periods[0]).toEqual('2015-01-01/2015-02-28')
+            expect(periods[periods.length - 1]).toEqual('2015-11-01/2015-12-31')
         })
     })
 })
