+++ src/period-calculation/get-fixed-period-by-date/index.ts
@@ -1,0 +1,1 @@
+export { default as getFixedPeriodByDate } from './get-fixed-period-by-date'
