+++ src/period-calculation/index.ts
@@ -1,1 +1,2 @@
-export { generateFixedPeriods } from './fixed-periods'
+export { generateFixedPeriods } from './generate-fixed-periods/index'
+export { getFixedPeriodByDate } from './get-fixed-period-by-date/index'
