--- src/period-calculation/monthly-periods/index.ts
@@ -1,2 +1,0 @@
-export { getMonthlyPeriods } from './get-monthly-periods'
-export { default as getMonthlyPeriodByDate } from './get-monthly-period-by-date'
