--- src/period-calculation/weekly-periods/index.ts
@@ -1,2 +1,0 @@
-export { getWeeklyPeriods } from './get-weekly-periods'
-export { default as getWeeklyPeriodByDate } from './get-weekly-period-by-date'
