--- src/period-calculation/yearly-periods/index.ts
@@ -1,2 +1,0 @@
-export { getYearlyPeriods } from './get-yearly-periods'
-export { default as getYearlyPeriodByDate } from './get-yearly-period-by-date'
